+++ src/compas_eve/__init__.py
@@ -17,6 +17,7 @@ Classes
     Topic
     Publisher
     Subscriber
+    EchoSubscriber
     Transport
     InMemoryTransport
     get_default_transport
@@ -36,7 +37,16 @@ __email__ = "casas@arch.ethz.ch"
 __version__ = "0.4.0"
 
 from .event_emitter import EventEmitterMixin  # noqa: F401 needed here to avoid circular import on py2.7
-from .core import Message, Publisher, Subscriber, Transport, Topic, get_default_transport, set_default_transport
+from .core import (
+    Message,
+    Publisher,
+    Subscriber,
+    EchoSubscriber,
+    Transport,
+    Topic,
+    get_default_transport,
+    set_default_transport,
+)
 from .memory import InMemoryTransport
 
 HERE = os.path.dirname(__file__)
@@ -49,6 +59,7 @@ __all__ = [
     "Message",
     "Publisher",
     "Subscriber",
+    "EchoSubscriber",
     "Topic",
     "Transport",
     "get_default_transport",